Moved storage package under services one
