+++ android/app/src/main/java/app/covidshield/services/storage/StorageDirectory.kt
@@ -1,4 +1,4 @@
-package app.covidshield.storage
+package app.covidshield.services.storage
 
 object StorageDirectory {
     val ExposureStatus: KeyDefinition = Unsecure ("exposureStatus")
