+++ android/app/src/main/java/app/covidshield/services/storage/StorageService.kt
@@ -1,4 +1,4 @@
-package app.covidshield.storage
+package app.covidshield.services.storage
 
 import android.content.Context
 import app.covidshield.utils.SingletonHolder
